feat: add swipe toggle to Tabs component and refactor Planner to use EventsTab content
